Refs #4993 Moving KeyedChanges to a separate file so that publisher and subscriber can share it
